Move to device_utils file
