+++ homeassistant/components/utility_meter/__init__.py
@@ -11,11 +11,7 @@ from homeassistant.components.sensor import DOMAIN as SENSOR_DOMAIN
 from homeassistant.config_entries import ConfigEntry
 from homeassistant.const import ATTR_ENTITY_ID, CONF_NAME, CONF_UNIQUE_ID, Platform
 from homeassistant.core import HomeAssistant, split_entity_id
-from homeassistant.helpers import (
-    device_registry as dr,
-    discovery,
-    entity_registry as er,
-)
+from homeassistant.helpers import device_utils as du, discovery, entity_registry as er
 import homeassistant.helpers.config_validation as cv
 from homeassistant.helpers.dispatcher import async_dispatcher_send
 from homeassistant.helpers.typing import ConfigType
@@ -192,7 +188,7 @@ async def async_setup(hass: HomeAssistant, config: ConfigType) -> bool:
 async def async_setup_entry(hass: HomeAssistant, entry: ConfigEntry) -> bool:
     """Set up Utility Meter from a config entry."""
 
-    await dr.async_remove_stale_device_links_helpers(
+    await du.async_remove_stale_device_links_helpers(
         hass, entry.entry_id, entry.options[CONF_SOURCE_SENSOR]
     )
